node-manager/internal-api/requirements.txt
beanie==1.18.0
fastapi==0.95.0
motor==3.1.2
PyJWT==2.6.0
pymongo==4.3.3
python-decouple==3.8
requests==2.28.2
uvicorn==0.21.1
azure-storage-blob
python-multipart
confluent_kafka
pymongo
pydantic
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
